+++ ragu/triplet/llm_artifact_extractor.py
@@ -11,7 +11,7 @@ from ragu.common.prompts.default_models import ArtifactsModel
 from ragu.common.prompts.prompt_storage import RAGUInstruction
 from ragu.common.prompts.messages import ChatMessages, render
 from ragu.graph.types import Entity, Relation
-from ragu.llm.llm import LLM
+from ragu.models.llm import LLM
 from ragu.triplet.base_artifact_extractor import BaseArtifactExtractor
 from ragu.triplet.types import NEREL_ENTITY_TYPES, NEREL_RELATION_TYPES
